Fix row selection
