--- react/components/EXPERIMENTAL_useCheckboxTree/README.md
@@ -1,6 +1,0 @@
-## Docs
-
-The `columns` property is an `Array` used to define the table columns and how they should behave visually. Each column describes each item field should be handled by the `Table`.
-
-```ts
-```
